Extract common docgen logic for later reuse
